uvat_rs

Module utils

Traits

MyIndex

Index trait encompassing operations from faer’s Index trait and some more.

Functions

apply_constraints

Apply constraints on matrix l and right hand side column vector rhs. Constraints are given by a list of indices c_indices and the values c_values : in the result, the values at index c_index[i] will equal c_value[i].

apply_constraints_mat

Same as apply_constraints but with a matrix for rhs and rows for constraints values.

build_edge_map

Build an edge map from faces f and total the number of vertices nv.

build_intrinsic_delaunay

Compute lengths, faces and edge map from the intrinsic Delaunay triangulations of the surface.

build_mass_cotan_laplacian_intrinsic

Build mass and stiffness matrices for the intrinsic Laplacian, from the intrinsic Delaunay triangulation (as precomputed by build_intrinsic_delaunay).

compute_tutte_parameterization

Compute Tutte parameterization with border b mapped to a circle.

flip_avoiding_line_search

Move the vertices from src by at most delta_dst while avoiding triangle flips and also minimizing the energy evaluted by e

get_boundary_loop

Find all ordered boundary loops from face indices f and edge map e.

is_delaunay

Returns wether the triangle pair with shaired edge l_s, shaire by triangle 1 ( l_s, l11, l12) and triangle 2 (l_s, l21, l22) is Delaunay.

map_to_circle

Compute len evenly space points on a circle

mollify

Add length to all edge lengths if the triangle inequality isn’t satisfied with enough margin for stability. From https://www.cs.cmu.edu/~kmcrane/Projects/NonmanifoldLaplace/NonmanifoldLaplace.pdf